modify: Revamped the method of setting up the environment.
