+++ library.properties
