+++ src/LovyanGFX.hpp
@@ -55,6 +55,7 @@ Contributors:
 
 #elif defined (__SAMD51__)
   #include "lgfx/platforms/lgfx_spi_samd51.hpp"
+
 //#elif defined (STM32F7)
 // not implemented.
 //  #include "lgfx/platforms/lgfx_spi_stm32_spi.hpp"
@@ -62,184 +63,43 @@ Contributors:
 #endif
 
 
+#if defined( LGFX_M5STACK ) || defined( ARDUINO_M5Stack_Core_ESP32 ) || defined( ARDUINO_M5STACK_FIRE ) // M5Stack
+  #include "config/LGFX_Config_M5Stack.hpp"
 
-namespace lgfx {
+#elif defined( LGFX_M5STICKC ) || defined( ARDUINO_M5Stick_C ) // M5Stick C
+  #include "config/LGFX_Config_M5StickC.hpp"
 
-#if defined( ARDUINO_M5Stack_Core_ESP32 ) || defined( ARDUINO_M5STACK_FIRE ) // M5Stack
+#elif defined( LGFX_ODROID_GO ) || defined( ARDUINO_ODROID_ESP32 ) // ODROID-GO
+  #include "config/LGFX_Config_ODROID_GO.hpp"
 
-  typedef Panel_M5Stack Panel_default;
-  struct LGFX_Config {
-    static constexpr spi_host_device_t spi_host = VSPI_HOST;
-    static constexpr int dma_channel = 1;
-    static constexpr int spi_mosi = 23;
-    static constexpr int spi_miso = 19;
-    static constexpr int spi_sclk = 18;
-  };
+#elif defined( LGFX_TTGO_TS ) || defined( ARDUINO_TTGO_T1 ) // TTGO TS
+  #include "config/LGFX_Config_TTGO_TS.hpp"
 
-#elif defined( ARDUINO_M5Stick_C ) // M5Stick C
+#elif defined( LGFX_TTGO_TWATCH ) || defined( ARDUINO_T ) // TTGO T-Watch
+  #include "config/LGFX_Config_TTGO_TWatch.hpp"
 
-  typedef Panel_M5StickC Panel_default;
-  struct LGFX_Config {
-    static constexpr spi_host_device_t spi_host = VSPI_HOST;
-    static constexpr int dma_channel = 1;
-    static constexpr int spi_mosi = 15;
-    static constexpr int spi_miso = 14;
-    static constexpr int spi_sclk = 13;
-  };
+#elif defined( LGFX_DDUINO32_XS ) || defined( ARDUINO_D ) || defined( ARDUINO_DDUINO32_XS )
+  #include "config/LGFX_Config_DDUINO32_XS.hpp"
 
-#elif defined( ARDUINO_ODROID_ESP32 ) // ODROID-GO
+#elif defined( LGFX_LOLIN_D32 ) || defined( ARDUINO_LOLIN_D32_PRO ) // LoLin D32 Pro
+  #include "config/LGFX_Config_LoLinD32.hpp"
 
-  typedef Panel_ODROID_GO Panel_default;
-  struct LGFX_Config {
-    static constexpr spi_host_device_t spi_host = VSPI_HOST;
-    static constexpr int dma_channel = 1;
-    static constexpr int spi_mosi = 23;
-    static constexpr int spi_miso = 19;
-    static constexpr int spi_sclk = 18;
-  };
+#elif defined( LGFX_ESP_WROVER_KIT ) || defined( ARDUINO_ESP32_DEV ) // ESP-WROVER-KIT
+  #include "config/LGFX_Config_ESP_WROVER_KIT.hpp"
 
-#elif defined( ARDUINO_TTGO_T1 ) // TTGO TS
+#elif defined( LGFX_WIO_TERMINAL ) || defined (ARDUINO_WIO_TERMINAL) || defined(WIO_TERMINAL)
+  #include "config/LGFX_Config_WioTerminal.hpp"
 
-  typedef Panel_TTGO_TS Panel_default;
-  struct LGFX_Config {
-    static constexpr spi_host_device_t spi_host = VSPI_HOST;
-    static constexpr int dma_channel = 1;
-    static constexpr int spi_mosi = 23;
-    static constexpr int spi_miso = -1;
-    static constexpr int spi_sclk =  5;
-  };
+#else
 
-#elif defined( ARDUINO_T ) // TTGO T-Watch
-
-  typedef Panel_TTGO_TWatch Panel_default;
-  struct LGFX_Config {
-    static constexpr spi_host_device_t spi_host = HSPI_HOST;
-    static constexpr int dma_channel = 1;
-    static constexpr int spi_mosi = 19;
-    static constexpr int spi_miso = -1;
-    static constexpr int spi_sclk = 18;
-  };
-
-#elif defined( ARDUINO_D ) || defined( ARDUINO_DDUINO32_XS )
-
-  typedef Panel_DDUINO32_XS Panel_default;
-  struct LGFX_Config {
-    static constexpr spi_host_device_t spi_host = VSPI_HOST;
-    static constexpr int dma_channel = 1;
-    static constexpr int spi_mosi = 26;
-    static constexpr int spi_miso = -1;
-    static constexpr int spi_sclk = 27;
-  };
-
-#elif defined( ARDUINO_LOLIN_D32_PRO ) // LoLin D32 Pro
-
-  typedef Panel_LoLinD32 Panel_default;
-  struct LGFX_Config {
-    static constexpr spi_host_device_t spi_host = VSPI_HOST;
-    static constexpr int dma_channel = 1;
-    static constexpr int spi_mosi = 23;
-    static constexpr int spi_miso = 19;
-    static constexpr int spi_sclk = 18;
-  };
-
-#elif defined( ARDUINO_ESP32_DEV ) // ESP-WROVER-KIT
-
-  struct Panel_default : public lgfx::Panel_ILI9341 {
-    Panel_default(void) {
-      spi_3wire = false;
-      spi_cs   = 22;
-      spi_dc   = 21;
-      gpio_rst = 18;
-      gpio_bl  = 5;
-      pwm_ch_bl = 7;
-      freq_write = 40000000;
-      freq_read  = 20000000;
-      freq_fill  = 80000000;
-      backlight_level = false;
-    }
-  };
-
-  struct LGFX_Config {
-    static constexpr spi_host_device_t spi_host = VSPI_HOST;
-    static constexpr int dma_channel = 1;
-    static constexpr int spi_mosi = 23;
-    static constexpr int spi_miso = 25;
-    static constexpr int spi_sclk = 19;
-  };
-
-#elif defined( ESP32 ) || ( CONFIG_IDF_TARGET_ESP32 ) // ESP-IDF ( or other panel )
-
-  typedef Panel_M5Stack Panel_default;
-
-  struct LGFX_Config {
-    static constexpr spi_host_device_t spi_host = VSPI_HOST;
-    static constexpr int dma_channel = 1;
-    static constexpr int spi_mosi = 23;
-    static constexpr int spi_miso = 19;
-    static constexpr int spi_sclk = 18;
-  };
-
-#elif defined (ARDUINO_WIO_TERMINAL) || defined(WIO_TERMINAL)
-
-  typedef Panel_WioTerminal Panel_default;
-
-  struct LGFX_Config {
-    static constexpr int sercom_index = 7;
-    static constexpr int sercom_clksrc = 0;   // -1=notchange / 0=select GCLK0
-    static constexpr int sercom_clkfreq = F_CPU;
-    static constexpr int spi_miso = 0x0112; // PORTB 18 (PORTB=0x0100 | 18=0x0012)
-    static constexpr int spi_mosi = 0x0113; // PORTB 19 (PORTB=0x0100 | 19=0x0013)
-    static constexpr int spi_sclk = 0x0114; // PORTB 20 (PORTB=0x0100 | 20=0x0014)
-    static constexpr SercomSpiTXPad pad_mosi = SPI_PAD_3_SCK_1;  // PAD_SPI3_TX;
-    static constexpr SercomRXPad    pad_miso = SERCOM_RX_PAD_2;  // PAD_SPI3_RX;
-  };
+  // If none of the above apply, Put a copy of "config/LGFX_Config" in this folder,
+  // and modify the content according to the environment.
+  // 上記のいずれにも該当しない場合、"config/LGFX_Config" のコピーをこのフォルダに配置し、
+  // 動作環境に応じて内容を修正してください。
+  #include "LGFX_Config.hpp"
 
 #endif
-}
 
-
-class LGFX : public lgfx::LGFX_SPI<lgfx::LGFX_Config>
-{
-public:
-  LGFX(void) : lgfx::LGFX_SPI<lgfx::LGFX_Config>()
-  {
-    static lgfx::Panel_default panel;
-
-    setPanel(&panel);
-  }
-
-#if defined( ARDUINO_ESP32_DEV ) // ESP-WROVER-KIT
-// ESP-WROVER-KIT is available in two types of panels. (ILI9341 or ST7789)
-  void initPanel(void) override {
-    if (!_panel) return;
-    _panel->init();
-
-    if (readPanelID() > 0) {  // check panel (ILI9341 or ST7789)
-      ESP_LOGI("LovyanGFX", "[Autodetect] Using Panel_ST7789");
-
-      static lgfx::Panel_ST7789 panel;
-      panel.spi_3wire = false;
-      panel.spi_cs   = 22;
-      panel.spi_dc   = 21;
-      panel.gpio_rst = 18;
-      panel.gpio_bl  = 5;
-      panel.pwm_ch_bl = 7;
-      panel.freq_write = 80000000;
-      panel.freq_read  = 16000000;
-      panel.freq_fill  = 80000000;
-      panel.backlight_level = false;
-      panel.spi_mode_read = 1;
-      panel.len_dummy_read_pixel = 16;
-
-      setPanel(&panel);
-    } else {
-      ESP_LOGI("LovyanGFX", "[Autodetect] Using Panel_ILI9341");
-    }
-    lgfx::LGFX_SPI<lgfx::LGFX_Config>::initPanel();
-  }
-#endif
-
-};
 
 
 #endif
