+++ src/config/LGFX_Config_TTGO_TWatch.hpp
@@ -1,0 +1,27 @@
+#ifndef LOVYANGFX_CONFIG_HPP_
+#define LOVYANGFX_CONFIG_HPP_
+
+namespace lgfx
+{
+  struct LGFX_Config
+  {
+    static constexpr spi_host_device_t spi_host = HSPI_HOST;
+    static constexpr int dma_channel = 1;
+    static constexpr int spi_mosi = 19;
+    static constexpr int spi_miso = -1;
+    static constexpr int spi_sclk = 18;
+  };
+}
+
+class LGFX : public lgfx::LGFX_SPI<lgfx::LGFX_Config>
+{
+public:
+  LGFX(void) : lgfx::LGFX_SPI<lgfx::LGFX_Config>()
+  {
+    static lgfx::Panel_TTGO_TWatch panel;
+
+    setPanel(&panel);
+  }
+};
+
+#endif
